订单服务-类型转换Bug解决【原因：远程调用order需要登录验证】

+++ gulimall-common/src/main/java/com/atguigu/common/to/mq/StockDetailTo.java
@@ -26,7 +26,7 @@ public class StockDetailTo {
     /**
      * 购买个数
      */
-    private Integer skuNum;
+    private Long skuNum;
     /**
      * 工作单id
      */
